Move common stages to stages/common
