+++ sgl-router/src/routers/grpc/stages/common/request_execution.rs
@@ -3,7 +3,7 @@
 use async_trait::async_trait;
 use axum::response::Response;
 
-use super::PipelineStage;
+use crate::routers::grpc::stages::PipelineStage;
 use crate::{
     grpc_client::{proto, sglang_scheduler::AbortOnDropStream},
     routers::grpc::{
